requirements.txt
aiobotocore==2.3.4
aiohttp==3.8.1
aioitertools==0.10.0
aiosignal==1.2.0
altair==4.2.0
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
asttokens==2.0.5
async-timeout==4.0.2
attrs==21.4.0
backcall==0.2.0
backoff==1.11.1
beautifulsoup4==4.11.1
bleach==5.0.0
blinker==1.4
blis==1.0.1
bokeh==2.4.3
boto3==1.21.21
botocore==1.24.21
branca==0.5.0
cachetools==5.2.0
catalogue==2.0.10
certifi==2020.4.5.1
cffi==1.15.0
charset-normalizer==2.0.12
click==8.0.4
cloudpathlib==0.20.0
cloudpickle==2.1.0
coiled==0.2.15
colorama==0.4.6
commonmark==0.9.1
confection==0.1.5
cycler==0.11.0
cymem==2.0.8
dask==2022.7.1
debugpy==1.6.0
decorator==5.1.1
defusedxml==0.7.1
distributed==2022.7.1
entrypoints==0.4
et-xmlfile==1.1.0
executing==0.8.3
fastjsonschema==2.15.3
filelock==3.14.0
Flask==2.1.3
folium==0.12.1.post1
fonttools==4.33.3
frozenlist==1.3.0
fsspec==2024.5.0
gitdb==4.0.11
GitPython==3.1.43
google-ai-generativelanguage==0.4.0
google-api-core==2.17.1
google-api-python-client==2.108.0
google-auth==2.23.4
google-auth-httplib2==0.1.1
google-auth-oauthlib==1.1.0
google-generativeai==0.4.1
googleapis-common-protos==1.63.0
grpcio==1.62.1
grpcio-status==1.62.1
HeapDict==1.0.1
htmlmin==0.1.12
httplib2==0.22.0
huggingface-hub==0.23.0
idna==2.6
ImageHash==4.2.1
importlib-metadata==4.11.4
intel-openmp==2021.4.0
ipykernel==6.13.0
ipython==8.4.0
ipython-genutils==0.2.0
ipywidgets==7.7.0
itsdangerous==2.1.2
jedi==0.18.1
Jinja2==3.1.2
jmespath==1.0.1
joblib==1.1.0
jsondiff==2.0.0
jsonschema==4.5.1
jupyter-client==7.3.1
jupyter-core==4.10.0
jupyterlab-pygments==0.2.2
jupyterlab-widgets==1.1.0
kiwisolver==1.4.2
langcodes==3.4.1
language_data==1.2.0
locket==1.0.0
marisa-trie==1.2.1
markdown-it-py==3.0.0
markdown2==2.4.13
MarkupSafe==2.1.1
matplotlib==3.5.2
matplotlib-inline==0.1.3
mdurl==0.1.2
missingno==0.5.1
mistune==0.8.4
mkl==2021.4.0
mpmath==1.3.0
msgpack==1.0.4
multidict==6.0.2
multimethod==1.8
murmurhash==1.0.10
nbclient==0.6.4
nbconvert==6.5.0
nbformat==5.4.0
nest-asyncio==1.5.5
networkx==2.8.4
nltk==3.8.1
notebook==6.4.11
numpy==1.22.4
oauth2client==4.1.3
openpyxl==3.1.2
packaging==21.3
pandas==2.2.2
pandas-profiling==3.2.0
pandocfilters==1.5.0
parso==0.8.3
partd==1.2.0
phik==0.12.2
pickleshare==0.7.5
pillow==10.3.0
plotly==5.8.0
preshed==3.0.9
prometheus-client==0.14.1
prompt-toolkit==3.0.29
proto-plus==1.23.0
protobuf==4.25.7
psutil==5.9.1
pure-eval==0.2.2
pyarrow==16.0.0
pyasn1==0.4.8
pyasn1-modules==0.2.1
pycparser==2.21
pydantic==1.9.1
pydeck==0.9.0
Pygments==2.17.2
Pympler==1.0.1
pyparsing==3.0.9
pyrsistent==0.18.1
python-dateutil==2.9.0.post0
python-decouple==3.8
pytz==2022.1
pytz-deprecation-shim==0.1.0.post0
PyWavelets==1.3.0
pywin32==304
pywinpty==2.0.5
pyxlsb==1.0.9
PyYAML==6.0
pyzmq==23.0.0
regex==2024.4.28
requests==2.31.0
rich==13.7.1
rsa==3.4.2
s3fs==2022.5.0
s3transfer==0.5.2
safetensors==0.4.3
scikit-learn==1.1.2
scipy==1.8.1
seaborn==0.11.2
semver==3.0.2
Send2Trash==1.8.0
shellingham==1.5.4
six==1.16.0
sklearn==0.0
smart-open==7.0.5
smmap==5.0.0
sortedcontainers==2.4.0
soupsieve==2.3.2.post1
spacy==3.8.2
spacy-legacy==3.0.12
spacy-loggers==1.0.5
srsly==2.4.8
stack-data==0.2.0
streamlit==1.33.0
streamlit-aggrid==1.0.5
streamlit-folium==0.6.14
streamlit-multipage==0.0.18
streamlit-pandas-profiling==0.1.3
sympy==1.12
tangled-up-in-unicode==0.2.0
tbb==2021.12.0
tblib==1.7.0
tenacity==8.2.3
terminado==0.15.0
textual==0.1.18
thinc==8.3.2
threadpoolctl==3.1.0
tinycss2==1.1.1
tokenizers==0.19.1
toml==0.10.2
toolz==0.11.2
torch==2.3.0
torchaudio==2.3.0
torchvision==0.18.0
tornado==6.1
tqdm==4.64.0
traitlets==5.2.2.post1
transformers==4.40.2
typer==0.12.5
typing_extensions==4.2.0
tzdata==2024.1
tzlocal==5.2
uritemplate==4.1.1
urllib3==1.25.9
validators==0.28.1
visions==0.7.4
wasabi==1.1.3
watchdog==2.1.8
wcwidth==0.2.5
weasel==0.4.1
webencodings==0.5.1
Werkzeug==2.1.2
widgetsnbextension==3.6.0
wincertstore==0.2
wordcloud==1.9.3
wrapt==1.14.1
XlsxWriter==3.0.3
yarl==1.7.2
zict==2.2.0
zipp==3.8.0
